+++ src/markdown/parsers/allParsers.mdx
@@ -3,4 +3,5 @@
 ### Want to convert some text to another format? Find useful parsers here
 
 #### CSV / TSV to Markdown
-- [CSV to GitHub Flavoured Markdown Table / MarkDown Bullet Points](/parsers)
+- [CSV to GitHub Flavoured Markdown Table ](/CSVTable)
+- [CSV to MarkDown Bullet Points](/CSVBullet)
